# services/api-gateway/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SBSMA - Scalable Backend System using Microservices API",
    "description": "Comprehensive REST API Documentation for the Scalable E-Commerce Backend System composed of 5 independent microservices.\n\n### Microservices Included:\n1. **User Service** (`:3001`)\n2. **Product Service** (`:3002`)\n3. **Order Service** (`:3003`)\n4. **Payment Service** (`:3004`)\n5. **Notification Service** (`:3005`)\n\nAll endpoints are accessible via this **API Gateway** (`:3000`).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway Server (Port 3000)"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users (Read - List)",
        "responses": {
          "200": { "description": "List of users retrieved successfully" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user (Create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Aarav Gupta" },
                  "email": { "type": "string", "example": "aarav@example.com" },
                  "password": { "type": "string", "example": "secret123" },
                  "role": { "type": "string", "example": "customer" },
                  "phone": { "type": "string", "example": "+91-9876500000" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Missing required fields" },
          "409": { "description": "User email already exists" }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID with addresses (Read - Single)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "User details retrieved" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user details (Update)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Aarav V. Gupta" },
                  "phone": { "type": "string", "example": "+91-9876599999" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated successfully" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by ID (Delete)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "User deleted successfully" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/users/{id}/addresses": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all addresses for a user",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "User addresses retrieved" } }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Add new delivery address for user",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["street", "city", "state", "zip_code"],
                "properties": {
                  "street": { "type": "string", "example": "100 Outer Ring Road" },
                  "city": { "type": "string", "example": "Bengaluru" },
                  "state": { "type": "string", "example": "Karnataka" },
                  "zip_code": { "type": "string", "example": "560103" },
                  "country": { "type": "string", "example": "India" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Address created successfully" } }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products with optional filters (Read - List)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search keyword" },
          { "name": "category_id", "in": "query", "schema": { "type": "integer" }, "description": "Category ID filter" }
        ],
        "responses": { "200": { "description": "Products retrieved successfully" } }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create new product (Create)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "price", "sku"],
                "properties": {
                  "category_id": { "type": "integer", "example": 1 },
                  "name": { "type": "string", "example": "Wireless Gaming Mouse" },
                  "description": { "type": "string", "example": "RGB Ultra-lightweight optical mouse" },
                  "price": { "type": "number", "example": 3499.00 },
                  "stock_quantity": { "type": "integer", "example": 100 },
                  "sku": { "type": "string", "example": "SKU-MOUSE-01" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Product created successfully" } }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID (Read - Single)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Product retrieved successfully" },
          "404": { "description": "Product not found" }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update product details (Update)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "price": { "type": "number", "example": 3299.00 },
                  "stock_quantity": { "type": "integer", "example": 90 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Product updated successfully" },
          "404": { "description": "Product not found" }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete product by ID (Delete)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Product deleted successfully" },
          "404": { "description": "Product not found" }
        }
      }
    },
...
  }
}